#406 - Fix the circular dependency issue between Polygon and Rectangle
