+++ packages/geometry/index.ts
@@ -1,4 +1,3 @@
 export * from './src/Point';
 export * from './src/Polygon';
-export * from './src/Rectangle';
 export * from './src/BezierCurve';
